requirements.txt
pandas
openpyxl
pdfplumber
langchain
langgraph
google-generativeai
reportlab
python-dotenv
